# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LamarIn API",
        "description": "REST API Sistem Informasi Lowongan Pekerjaan — Proyek Akhir Kelompok",
        "contact": {
            "email": "kelompok@lamarIn.dev"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Server Utama"
        }
    ],
    "paths": {
...
        "/api/lowongan": {
            "get": {
                "tags": [
                    "Lowongan"
                ],
                "summary": "Melihat lowongan (penyedia: milik sendiri | pelamar: semua aktif)",
                "operationId": "d069a3f898898cb8c4fbc5ef3112e190",
                "responses": {
                    "200": {
                        "description": "Daftar lowongan berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Lowongan"
                ],
                "summary": "Membuat lowongan baru (hanya Penyedia)",
                "operationId": "1f93d81137336371123964222d9b255e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "judul",
                                    "deskripsi",
                                    "kualifikasi",
                                    "lokasi",
                                    "jenis_pekerjaan",
                                    "batas_daftar"
                                ],
                                "properties": {
                                    "judul": {
                                        "type": "string",
                                        "example": "Backend Developer"
                                    },
                                    "deskripsi": {
                                        "type": "string",
                                        "example": "Membangun REST API"
                                    },
                                    "kualifikasi": {
                                        "type": "string",
                                        "example": "Min. S1 Informatika"
                                    },
                                    "lokasi": {
                                        "type": "string",
                                        "example": "Malang"
                                    },
                                    "jenis_pekerjaan": {
                                        "type": "string",
                                        "enum": [
                                            "full-time",
                                            "part-time",
                                            "remote",
                                            "kontrak"
                                        ]
                                    },
                                    "gaji_min": {
                                        "type": "integer",
                                        "example": 5000000
                                    },
                                    "gaji_max": {
                                        "type": "integer",
                                        "example": 10000000
                                    },
                                    "batas_daftar": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-06-30"
                                    },
                                    "kategori_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Lowongan berhasil dibuat"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden — bukan penyedia"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/lowongan/{id}": {
            "get": {
                "tags": [
                    "Lowongan"
                ],
                "summary": "Melihat detail satu lowongan",
                "operationId": "0afcdc2beb816cd2af76787c43354672",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail lowongan"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Lowongan tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Lowongan"
                ],
                "summary": "Memperbarui informasi lowongan",
                "operationId": "579fdfe0a3004e1dcd92aab4a619269c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "judul": {
                                        "type": "string"
                                    },
                                    "deskripsi": {
                                        "type": "string"
                                    },
                                    "kualifikasi": {
                                        "type": "string"
                                    },
                                    "lokasi": {
                                        "type": "string"
                                    },
                                    "jenis_pekerjaan": {
                                        "type": "string",
                                        "enum": [
                                            "full-time",
                                            "part-time",
                                            "remote",
                                            "kontrak"
                                        ]
                                    },
                                    "gaji_min": {
                                        "type": "integer"
                                    },
                                    "gaji_max": {
                                        "type": "integer"
                                    },
                                    "batas_daftar": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "kategori_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lowongan berhasil diperbarui"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Lowongan tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Lowongan"
                ],
                "summary": "Menghapus lowongan",
                "operationId": "fa3332b068d0ad4f13299d5c29054ca7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lowongan berhasil dihapus"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Lowongan tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/lowongan/{id}/status": {
            "patch": {
                "tags": [
                    "Lowongan"
                ],
                "summary": "Mengubah status aktif/nonaktif lowongan",
                "operationId": "e7be6114cea2ea4b964bb8e19d7d91b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "aktif",
                                            "nonaktif"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status lowongan berhasil diubah"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Lowongan tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}